move the logic of handling when a message is sent to MessageSentHandler
